block: Make the rwbs field as a bit field enum
The rwbs value of block request event is in fact a series of bit fields
set to 0 or 1. The enumeration values are all powers of 2, trace readers
should interpret this as a bit field enum and show the result as a
concatenation of the corresponding flags. For example, a rwbs value of
12 means flag 4 and 8 are active and the result should be
RWBS_FLAG_READ|RWBS_FLAG_RAHEAD.

Signed-off-by: Geneviève Bastien <gbastien@versatic.net>