tracepoint: Refactor representation of nested types
Refactor enumeration, sequence, array, structure, and variant types.

Implement internal data structures to support nested types.

All probe providers using ctf_enum(), ctf_array*() and ctf_sequence*()
are switched to this new internal type representation.

Each of sequence, array, struct and variant gain a "alignment" property,
which is a feature which was needed in lttng-modules to express
alignment for an array or sequence of bits.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>