Cleanup: Rename patches.i to patches.h
This generated header file contains a list of patches applied on the
lttng-modules tree. Based on the C99 specification, ".i" files are not
supposed to be preprocessed, although this header header file is
expected to be preprocessed.

Rename it from ".i" to ".h" to convey that it is a C header meant to be
preprocessed.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>